workload/anns/faiss-pq-system/requirements.txt
faiss-cpu>=1.7.0
numpy>=1.21.0
pandas>=1.3.0
PyYAML>=6.0
scikit-learn>=1.0.0
tqdm>=4.62.0
requests>=2.26.0
h5py>=3.6.0
matplotlib>=3.5.0
seaborn>=0.11.0 